requirements.txt
discord.py==2.2.3
python-dotenv==1.0.0
requests==2.30.0
bs4==0.0.1
APScheduler==3.10.1